add erro delete residencia

+++ ListMEAPI/ListMEAPI/Repositories/ResidenciaRepository.cs
@@ -16,6 +16,7 @@ namespace ListMEAPI.Repositories
             _context = ctx;
         }
 
+        //POST
         public dynamic Create(ResidenciaModel residencia)
         {
             _context.Add(residencia);
@@ -24,31 +25,44 @@ namespace ListMEAPI.Repositories
             return null;
         }
 
+        //GET
         public List<ResidenciaModel> GetAll()
         {
             return _context.Residencias.ToList();
         }
 
+        //GET USUÁRIO POR ID
         public UsuarioModel GetUsuario(int Id)
         {
             var usuarioRequerido = _context.Usuarios.Find(Id);
             return usuarioRequerido;
         }
 
+        //GET RESIDENCIA POR ID
         public ResidenciaModel GetOneResidencia(int Id)
         {
             var residenciaRequerida = _context.Residencias.Find(Id);
             return residenciaRequerida;
         }
 
-        public void Delete(int id)
+        //DELETE
+        public bool Delete(int id)
         {
             var result = _context.Residencias.Find(id);
-            _context.Residencias.Remove(result);
-            _context.SaveChanges();
+            if (result != null)
+            {
+                _context.Residencias.Remove(result);
+                _context.SaveChanges();
+                return true;
+            }
+            else
+            {
+                return false;
+            }
 
         }
 
+        //PUT
         public ResidenciaModel Update(int Id, CadastroResidenciaRequest residenciaAtualizada)
         {
             ResidenciaModel residenciaAntiga = _context.Residencias.Find(Id);
